Refactor navigation drawer - remove pagination - disabled expanding on hover - add tooltips in mini-variant mode - move to ui/layout
