+++ src/stores/layout.js
@@ -3,17 +3,15 @@ import { ref, computed, getCurrentInstance } from 'vue'
 import panelViewConfig from '@/assets/panelViewConfig.json'
 
 const generateDefaultConfig = () =>
-    Array.from({ length: 3 }, () =>
-        Object.fromEntries(
-            panelViewConfig.map(({ name }) => [
-                name,
-                {
-                    shape: [],
-                    windows: {},
-                    nextId: 0,
-                },
-            ])
-        )
+    Object.fromEntries(
+        panelViewConfig.map(({ name }) => [
+            name,
+            {
+                shape: [],
+                windows: {},
+                nextId: 0,
+            },
+        ])
     )
 
 export const useLayoutStore = defineStore('layout', () => {
@@ -39,19 +37,14 @@ export const useLayoutStore = defineStore('layout', () => {
         const $route = getCurrentInstance().proxy.$route
         return $route.params.variant
     })
-    const screen = computed(() => {
-        const $route = getCurrentInstance().proxy.$route
-        return $route.params.screen
-    })
     const layout = computed(() => {
-        return allLayouts.value[screen.value - 1][panel.value]
+        return allLayouts.value[panel.value]
     })
     function nextWindowId() {
-        return `${screen.value};${panel.value};${layout.value.nextId++}`
+        return `${panel.value};${layout.value.nextId++}`
     }
 
     return {
-        screen,
         panel,
         allLayouts,
         layout,
