+++ src/views/PanelView.vue
@@ -3,7 +3,7 @@ import { computed } from 'vue'
 import { GridLayout, GridItem } from 'vue-grid-layout'
 
 import AppBar from '@/components/AppBar'
-import NavigationDrawer from '@/components/NavigationDrawer.vue'
+import NavigationDrawer from '@/ui/layout/NavigationDrawer.vue'
 import EditDrawer from '@/components/EditDrawer'
 import BatteryBar from '@/components/BatteryBar.vue'
